+++ src/map/skills/mercenary/mercenary_bash.cpp
@@ -12,7 +12,7 @@ void SkillMercenaryBash::calculateSkillRatio(const Damage* wd, const block_list*
 	base_skillratio += 30 * skill_lv;
 }
 
-void SkillMercenaryBash::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillMercenaryBash::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	// +5% hit per level
 	hit_rate += hit_rate * 5 * skill_lv / 100;
 }
