+++ src/map/skills/mercenary/mercenary_magnumbreak.cpp
@@ -44,6 +44,6 @@ void SkillMercenaryMagnumBreak::castendNoDamageId(block_list *src, block_list *t
 #endif
 }
 
-void SkillMercenaryMagnumBreak::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillMercenaryMagnumBreak::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 10 * skill_lv / 100;
 }
