+++ src/map/skills/mercenary/mercenary_pierce.cpp
@@ -18,6 +18,6 @@ void SkillMercenaryPierce::calculateSkillRatio(const Damage *wd, const block_lis
 	base_skillratio += 10 * skill_lv;
 }
 
-void SkillMercenaryPierce::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillMercenaryPierce::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 5 * skill_lv / 100;
 }
