+++ src/map/skills/merchant/cartrevolution.hpp
@@ -12,6 +12,6 @@ public:
 	SkillCartRevolution();
 
 	void calculateSkillRatio(const Damage* wd, const block_list* src, const block_list* target, uint16 skill_lv, int32& base_skillratio, int32 mflag) const override;
-	void modifyHitRate(int16& hit_rate, const block_list *src, const block_list *target, uint16 skill_lv) const override;
+	void modifyHitRate(int32& hit_rate, const block_list *src, const block_list *target, uint16 skill_lv) const override;
 	void castendDamageId(block_list* src, block_list* target, uint16 skill_lv, t_tick tick, int32& flag) const override;
 };
