+++ src/map/skills/merchant/carttornado.cpp
@@ -27,7 +27,7 @@ void SkillCartTornado::castendNoDamageId(block_list *src, block_list *target, ui
 	skill_castend_damage_id(src, src, getSkillId(), skill_lv, tick, flag);
 }
 
-void SkillCartTornado::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillCartTornado::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	const map_session_data* sd = BL_CAST(BL_PC, src);
 
 	if (sd && pc_checkskill(sd, GN_REMODELING_CART))
