+++ src/map/skills/npc/bleeding2.cpp
@@ -12,6 +12,6 @@ void SkillBleeding2::applyAdditionalEffects(block_list *src, block_list *target,
 	sc_start(src,target,SC_BLEEDING,(50+10*skill_lv),skill_lv,skill_get_time2(getSkillId(),skill_lv));
 }
 
-void SkillBleeding2::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillBleeding2::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
