+++ src/map/skills/npc/darknessbreath.cpp
@@ -31,6 +31,6 @@ void SkillDarknessBreath::castendDamageId(block_list *src, block_list *target, u
 	}
 }
 
-void SkillDarknessBreath::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillDarknessBreath::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate *= 2;
 }
