+++ src/map/skills/npc/earthattributeattack.cpp
@@ -10,6 +10,6 @@ void SkillEarthAttributeAttack::calculateSkillRatio(const Damage *wd, const bloc
 	base_skillratio += 100 * (skill_lv - 1);
 }
 
-void SkillEarthAttributeAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillEarthAttributeAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
