+++ src/map/skills/npc/firebreath.cpp
@@ -31,6 +31,6 @@ void SkillFireBreath::castendDamageId(block_list *src, block_list *target, uint1
 	}
 }
 
-void SkillFireBreath::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillFireBreath::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate *= 2;
 }
