+++ src/map/skills/npc/holyattributeattack.cpp
@@ -10,6 +10,6 @@ void SkillHolyAttributeAttack::calculateSkillRatio(const Damage *wd, const block
 	base_skillratio += 100 * (skill_lv - 1);
 }
 
-void SkillHolyAttributeAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillHolyAttributeAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
