+++ src/map/skills/npc/icebreath.cpp
@@ -36,6 +36,6 @@ void SkillIceBreath::castendDamageId(block_list *src, block_list *target, uint16
 	}
 }
 
-void SkillIceBreath::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillIceBreath::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate *= 2;
 }
