+++ src/map/skills/npc/icebreath2.cpp
@@ -36,6 +36,6 @@ void SkillIceBreath2::castendDamageId(block_list *src, block_list *target, uint1
 	}
 }
 
-void SkillIceBreath2::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillIceBreath2::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate *= 2;
 }
