+++ src/map/skills/npc/petrifyattack.cpp
@@ -12,6 +12,6 @@ void SkillPetrifyAttack::applyAdditionalEffects(block_list *src, block_list *tar
 	sc_start2(src,target,SC_STONEWAIT,(20*skill_lv),skill_lv,src->id,skill_get_time2(getSkillId(),skill_lv),skill_get_time(getSkillId(), skill_lv));
 }
 
-void SkillPetrifyAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillPetrifyAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
