+++ src/map/skills/npc/poisonattack.cpp
@@ -12,6 +12,6 @@ void SkillPoisonAttack::applyAdditionalEffects(block_list *src, block_list *targ
 	sc_start(src,target,SC_POISON,(20*skill_lv),skill_lv,skill_get_time2(getSkillId(),skill_lv));
 }
 
-void SkillPoisonAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillPoisonAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
