+++ src/map/skills/npc/poisonattributeattack.cpp
@@ -10,6 +10,6 @@ void SkillPoisonAttributeAttack::calculateSkillRatio(const Damage *wd, const blo
 	base_skillratio += 100 * (skill_lv - 1);
 }
 
-void SkillPoisonAttributeAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillPoisonAttributeAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
