+++ src/map/skills/npc/silenceattack.cpp
@@ -12,6 +12,6 @@ void SkillSilenceAttack::applyAdditionalEffects(block_list *src, block_list *tar
 	sc_start(src,target,SC_SILENCE,(20*skill_lv),skill_lv,skill_get_time2(getSkillId(),skill_lv));
 }
 
-void SkillSilenceAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillSilenceAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
