+++ src/map/skills/npc/stunattack.cpp
@@ -12,6 +12,6 @@ void SkillStunAttack::applyAdditionalEffects(block_list *src, block_list *target
 	sc_start(src,target,SC_STUN,(20*skill_lv),skill_lv,skill_get_time2(getSkillId(),skill_lv));
 }
 
-void SkillStunAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillStunAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
