+++ src/map/skills/npc/undeadattributechange.cpp
@@ -12,6 +12,6 @@ void SkillUndeadAttributeChange::applyAdditionalEffects(block_list *src, block_l
 	sc_start(src, target, SC_CHANGEUNDEAD, (10 * skill_lv), skill_lv, skill_get_time2(getSkillId(), skill_lv));
 }
 
-void SkillUndeadAttributeChange::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillUndeadAttributeChange::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
