+++ src/map/skills/npc/undeadelementattack.cpp
@@ -10,6 +10,6 @@ void SkillUndeadElementAttack::calculateSkillRatio(const Damage *wd, const block
 	base_skillratio += 100 * (skill_lv - 1);
 }
 
-void SkillUndeadElementAttack::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillUndeadElementAttack::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 20 / 100;
 }
