+++ src/map/skills/npc/windattributeattack.hpp
@@ -10,5 +10,5 @@ public:
 	SkillWindAttributeAttack();
 
 	void calculateSkillRatio(const Damage *wd, const block_list *src, const block_list *target, uint16 skill_lv, int32 &base_skillratio, int32 mflag) const override;
-	void modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const override;
+	void modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const override;
 };
