+++ src/map/skills/swordman/sonicwave.cpp
@@ -15,6 +15,6 @@ void SkillSonicWave::calculateSkillRatio(const Damage* wd, const block_list* src
 	RE_LVL_DMOD(100);
 }
 
-void SkillSonicWave::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillSonicWave::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	hit_rate += hit_rate * 3 * skill_lv / 100; // !TODO: Confirm the hitrate bonus
 }
