+++ src/map/skills/thief/sonicblow.cpp
@@ -36,7 +36,7 @@ void SkillSonicBlow::applyAdditionalEffects(block_list *src, block_list *target,
 		sc_start(src, target, SC_STUN, (2 * skill_lv + 10), skill_lv, skill_get_time2(getSkillId(), skill_lv));
 }
 
-void SkillSonicBlow::modifyHitRate(int16& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
+void SkillSonicBlow::modifyHitRate(int32& hit_rate, const block_list* src, const block_list* target, uint16 skill_lv) const {
 	const map_session_data* sd = BL_CAST( BL_PC, src );
 
 	if(sd && pc_checkskill(sd,AS_SONICACCEL) > 0)
